Follow MS-LSAD and MS-LSAT spec for LSARPC & LookupSids
