+++ lib/msf/core/exploit/remote/ms_lsad.rb
@@ -6,7 +6,7 @@
 
 module Msf
 
-  module Exploit::Remote::MsLsarpc
+  module Exploit::Remote::MsLsad
 
     include Msf::Exploit::Remote::SMB::Client::Ipc
 
@@ -66,16 +66,6 @@ module Msf
       self.lsarpc_pipe.lsar_query_information_policy(
         policy_handle: policy_handle,
         information_class: information_class
-      )
-    end
-
-    def lookup_sids(policy_handle, sids, lookup_level)
-      sids = [sids] unless sids.is_a?(Array)
-
-      self.lsarpc_pipe.lsar_lookup_sids(
-        policy_handle: policy_handle,
-        sids: sids,
-        lookup_level: lookup_level
       )
     end
